backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pyyaml==6.0.2
neo4j==5.25.0
pydantic==2.9.2
cryptography>=43.0.0
openai>=1.40.0
